python_api/app-6901/requirements.txt
fastapi
uvicorn
python-multipart
requests>=2.31.0
yt-dlp
edge-tts
openai-whisper
torch
torchvision
numpy
pillow
timm
kornia
scikit-image
scikit-learn
huggingface_hub>=0.23.0
transformers>=4.40.0
opencv-python
accelerate>=0.27.0
tokenizers>=0.22.0
safetensors>=0.4.3
sentencepiece>=0.2.0
selenium>=4.15.0
ddgs
python-dotenv
keyring>=25.0.0
beautifulsoup4
lxml
webdriver_manager
undetected-chromedriver>=3.5.5
setuptools
